Set root API URL in the Settings app; other fixes
This allows me to publish the app to the App Store and let others point
it to their locally-hosted instances of the server. Note that storing
the settings in the Settings app means they are stored in UserDefaults,
which is why I use `@AppStorage` to access them.

Other changes include:
- Providing a button to jump to settings when the root API URL is unset
- update the API to expect a CreateFileResult upon creating new files,
  instead of the FileSummary that it used to return. I also provided a
  convenience function to convert CreateFileResult to FileSummary
- Support branches named 'master' as well as 'main'
- Allow insecure (i.e., HTTP) connections, trusting that users will only
  use them over a VPN